+++ source/platform/nx/common.h
@@ -195,9 +195,9 @@ int MSG_ReadControlHeader(void);
 
 //============================================================================
 
-void Q_memset (void *dest, int fill, int count);
-void Q_memcpy (void *dest, void *src, int count);
-int Q_memcmp (void *m1, void *m2, int count);
+void Q_memset (void *dest, int fill, size_t count);
+void Q_memcpy (void *dest, void *src, size_t count);
+int Q_memcmp (void *m1, void *m2, size_t count);
 void Q_strcpy (char *dest, char *src);
 void Q_strncpy (char *dest, char *src, int count);
 int Q_strlen (char *str);
